+++ src/basilisp/lang/compiler/optimizer.py
@@ -5,10 +5,11 @@ import basilisp._pyast as ast
 
 def _filter_dead_code(nodes: Iterable[ast.AST]) -> List[ast.AST]:
     """Return a list of body nodes, trimming out unreachable code (any
-    statements appearing after `break`, `continue`, and `return` nodes)."""
+    statements appearing after `break`, `continue`, `raise`, and `return`
+    nodes)."""
     new_nodes: List[ast.AST] = []
     for node in nodes:
-        if isinstance(node, (ast.Break, ast.Continue, ast.Return)):
+        if isinstance(node, (ast.Break, ast.Continue, ast.Raise, ast.Return)):
             new_nodes.append(node)
             break
         new_nodes.append(node)
@@ -52,17 +53,31 @@ class PythonASTOptimizer(ast.NodeTransformer):
         )
 
     def visit_If(self, node: ast.If) -> Optional[ast.AST]:
-        """Eliminate dead code from if/elif bodies."""
+        """Eliminate dead code from if/elif bodies.
+
+        If the new `if` statement `body` is empty after eliminating dead code, replace
+        the body with the `orelse` body and negate the `if` condition.
+
+        If both the `body` and `orelse` body are empty, eliminate the node from the
+        tree."""
         new_node = self.generic_visit(node)
         assert isinstance(new_node, ast.If)
-        return ast.copy_location(
-            ast.If(
-                test=new_node.test,
-                body=_filter_dead_code(new_node.body),
-                orelse=_filter_dead_code(new_node.orelse),
-            ),
-            new_node,
-        )
+
+        new_body = _filter_dead_code(new_node.body)
+        new_orelse = _filter_dead_code(new_node.orelse)
+
+        if new_body:
+            ifstmt = ast.If(test=new_node.test, body=new_body, orelse=new_orelse,)
+        elif new_orelse:
+            ifstmt = ast.If(
+                test=ast.UnaryOp(op=ast.Not(), operand=new_node.test),
+                body=new_orelse,
+                orelse=[],
+            )
+        else:
+            return None
+
+        return ast.copy_location(ifstmt, new_node)
 
     def visit_While(self, node: ast.While) -> Optional[ast.AST]:
         """Eliminate dead code from while bodies."""